Added to day class, made a function for sell lemonade that loops throught the customer list if they bought or not. Working on subtracting a cup and adding money if they did.

+++ UserInterface.cs
@@ -81,7 +81,7 @@ namespace LemonadeStand_3DayStarter
         public static double PricePerCup(Inventory inventory)
         {
             Console.WriteLine("Please enter your price per cup!");
-            double pricePerCup = Convert.ToInt32(Console.ReadLine().Trim());
+            double pricePerCup = Convert.ToDouble(Console.ReadLine().Trim());
             return pricePerCup;
         }
